refactor: remove unused keysigning components to streamline codebase
- Deleted the useJoinKeysignUrlQuery hook and associated components to enhance code clarity and reduce complexity.
- Removed DownloadKeysignQrCode and KeysignPeerDiscoveryStep components, which were dependent on the deleted hook.
- Cleaned up imports across affected files to maintain organization and improve maintainability.
