+++ clients/desktop/src/vault/keysign/start/StartSecureKeysignFlow.tsx
@@ -1,10 +1,10 @@
+import { KeysignPeerDiscoveryStep } from '@core/ui/mpc/keysign/peers/KeysignPeerDiscoveryStep'
 import { StartMpcSessionFlow } from '@core/ui/mpc/session/StartMpcSessionFlow'
 import { MpcPeersSelectionProvider } from '@core/ui/mpc/state/mpcSelectedPeers'
 import { useCorePathState } from '@core/ui/navigation/hooks/useCorePathState'
 import { StepTransition } from '@lib/ui/base/StepTransition'
 
 import { KeysignSigningStep } from '../shared/KeysignSigningStep'
-import { KeysignPeerDiscoveryStep } from './peerDiscovery/KeysignPeerDiscoveryStep'
 
 export const StartSecureKeysignFlow = () => {
   const { keysignPayload } = useCorePathState<'keysign'>()
