+++ core/ui/mpc/keysign/DownloadKeysignQrCode.tsx
@@ -1,4 +1,5 @@
 import { SaveAsImage } from '@core/ui/file/SaveAsImage'
+import { useJoinKeysignUrlQuery } from '@core/ui/mpc/keysign/queries/useJoinKeysignUrlQuery'
 import { PrintableQrCode } from '@core/ui/qr/PrintableQrCode'
 import { getVaultPublicKeyExport } from '@core/ui/vault/share/utils/getVaultPublicKeyExport'
 import { useCurrentVault } from '@core/ui/vault/state/currentVault'
@@ -6,8 +7,6 @@ import { FileUpIcon } from '@lib/ui/icons/FileUpIcon'
 import { PageHeaderIconButton } from '@lib/ui/page/PageHeaderIconButton'
 import { MatchQuery } from '@lib/ui/query/components/MatchQuery'
 import { useTranslation } from 'react-i18next'
-
-import { useJoinKeysignUrlQuery } from '../../shared/queries/useJoinKeysignUrlQuery'
 
 export const DownloadKeysignQrCode = () => {
   const joinKeysignUrlQuery = useJoinKeysignUrlQuery()
