+++ core/ui/mpc/keysign/peers/KeysignPeerDiscoveryStep.tsx
@@ -9,7 +9,10 @@ import { PeersManagerFrame } from '@core/ui/mpc/devices/peers/PeersManagerFrame'
 import { PeersManagerTitle } from '@core/ui/mpc/devices/peers/PeersManagerTitle'
 import { PeersPageContentFrame } from '@core/ui/mpc/devices/peers/PeersPageContentFrame'
 import { useMpcPeerOptionsQuery } from '@core/ui/mpc/devices/queries/useMpcPeerOptionsQuery'
+import { DownloadKeysignQrCode } from '@core/ui/mpc/keysign/DownloadKeysignQrCode'
+import { useJoinKeysignUrlQuery } from '@core/ui/mpc/keysign/queries/useJoinKeysignUrlQuery'
 import { MpcLocalServerIndicator } from '@core/ui/mpc/server/MpcLocalServerIndicator'
+import { useMpcDevices } from '@core/ui/mpc/state/mpcDevices'
 import { useMpcServerType } from '@core/ui/mpc/state/mpcServerType'
 import { useCurrentVault } from '@core/ui/vault/state/currentVault'
 import { getFormProps } from '@lib/ui/form/utils/getFormProps'
@@ -22,17 +25,22 @@ import { OnFinishProp } from '@lib/ui/props'
 import { QueryBasedQrCode } from '@lib/ui/qr/QueryBasedQrCode'
 import { MatchQuery } from '@lib/ui/query/components/MatchQuery'
 import { range } from '@lib/utils/array/range'
-import { useEffect } from 'react'
+import { useEffect, useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
-
-import { useJoinKeysignUrlQuery } from '../../shared/queries/useJoinKeysignUrlQuery'
-import { DownloadKeysignQrCode } from './DownloadKeysignQrCode'
-import { useIsPeerDiscoveryStepDisabled } from './hooks/useIsPeerDiscoveryStepDisabled'
 
 export const KeysignPeerDiscoveryStep = ({ onFinish }: OnFinishProp) => {
   const { t } = useTranslation()
 
-  const isDisabled = useIsPeerDiscoveryStepDisabled()
+  const devices = useMpcDevices()
+  const { signers } = useCurrentVault()
+
+  const isDisabled = useMemo(() => {
+    const requiredDevicesNumber = getKeygenThreshold(signers.length)
+
+    if (devices.length !== requiredDevicesNumber) {
+      return t('select_n_devices', { count: requiredDevicesNumber - 1 })
+    }
+  }, [devices.length, signers.length, t])
 
   useEffect(() => {
     if (!isDisabled) {
@@ -43,8 +51,6 @@ export const KeysignPeerDiscoveryStep = ({ onFinish }: OnFinishProp) => {
   const joinUrlQuery = useJoinKeysignUrlQuery()
 
   const [serverType] = useMpcServerType()
-
-  const { signers } = useCurrentVault()
 
   const requiredSigners = getKeygenThreshold(signers.length)
   const requiredPeers = requiredSigners - 1
